feat: render chat text/glow colors as raw RGBA instead of UIColor rows
Both vanilla chat and Chat 2 support arbitrary RGBA colors via raw SeString
Color/EdgeColor macros, not just the FFXIV UIColor palette, so Text and Glow
color pickers across Formatting and Groups now use a plain RGBA picker
instead of the old named-palette swatch grid. Alpha stays hidden for
Text/Glow (vanilla ignores it there) and only stays editable for Chat 2
group backgrounds, a separate system that already supported it.

Range-fade dimming now darkens this plugin's own colors via direct RGB math
instead of nearest-UIColor-row search; pre-existing game-embedded UIColor
payloads (e.g. item/map links) still dim through the old row-based path.

No config migration: existing ushort UIColor-row values will silently
reinterpret as different RGBA colors on load, accepted since the plugin
isn't on the official repo yet.

+++ GobchatEx/Config/PlayerGroup.cs
@@ -8,7 +8,8 @@ namespace GobchatEx.Config;
 /// One player group (Milestone 2): either a custom group matched by a <see cref="Members"/> list, or
 /// one of the game's seven fixed friend-list display groups matched by <see cref="FfGroup"/>
 /// (0=Star..6=Club). Sender-name recoloring uses <see cref="Foreground"/>/<see cref="Glow"/> the same
-/// way segment styles do; 0 means "do not recolor".
+/// way segment styles do — packed RGBA (0xRRGGBBAA), same format as <see cref="ChatTwoBackground"/>,
+/// rendered via a raw SeString Color/EdgeColor macro; 0 means "do not recolor".
 /// </summary>
 [Serializable]
 public class PlayerGroup
@@ -25,8 +26,8 @@ public class PlayerGroup
     /// </summary>
     public List<GroupMember> Members { get; set; } = [];
 
-    public ushort Foreground { get; set; }
-    public ushort Glow { get; set; }
+    public uint Foreground { get; set; }
+    public uint Glow { get; set; }
 
     /// <summary>
     /// Per-message background color rendered in Chat 2 (0xRRGGBBAA, 0 = none). Chat 2-only: the
